move WithCachedTypeInfo to rustc_type_ir
